requirements.txt
numpy==1.24.4
opencv_python==4.8.0.74
Pillow==10.0.0
torch_summary==1.4.5
tqdm==4.65.0
tensorboard
matplotlib
gradio
typing_extensions